pyproject.toml
[project]
name = "sigmon"
description = "sigsum log monitor"
license = "Apache-2.0"
readme = "README.md"
authors = [
    { name = "Florian Larysch", email = "fl@n621.de" }
]
dynamic = ["version"]
requires-python = ">=3.11"
dependencies = [
 "pynacl>=1.5.0",
 "requests>=2.32.4",
]

[project.scripts]
sigmon = "sigmon.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/sigmon/__about__.py"
